added float

+++ src/Main.java
@@ -4,9 +4,9 @@ public class Main {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Enter first value");
-        int x = scanner.nextInt();
+        float x = scanner.nextFloat();
         System.out.println("Enter second value");
-        int y = scanner.nextInt();
+        float y = scanner.nextFloat();
         System.out.println("What do you want to do?\nSelect:\n1.addition\n2.substraction\n3.multiplication\n4.division");
         int select = scanner.nextInt();
         switch (select) {
